[#440] Refactor SceneAcceptor
Extract a SceneAcceptor interface and compose the object
using decorator, allowing for later extensions.
